Automatically initialize configuration
No longer require running `skein init` before anything else. Now
defaults will be written (with a warning) on the first call
automatically.
